Cleanup, optimizations and small improvements

+++ data/chat/src/main/java/se/onemanstudio/playaroundwithai/data/chat/data/local/dao/PromptsHistoryDao.kt
@@ -19,23 +19,23 @@ interface PromptsHistoryDao {
     suspend fun getPromptsBySyncStatus(status: String): List<PromptEntity>
 
     @Query("UPDATE prompt_history SET syncStatus = :status WHERE id = :id")
-    suspend fun updateSyncStatus(id: Int, status: String)
+    suspend fun updateSyncStatus(id: Long, status: String)
 
     @Query("SELECT COUNT(*) FROM prompt_history WHERE syncStatus = :status")
     fun getCountBySyncStatus(status: String): Flow<Int>
 
     @Query("UPDATE prompt_history SET text = :text WHERE id = :id")
-    suspend fun updatePromptText(id: Int, text: String)
+    suspend fun updatePromptText(id: Long, text: String)
 
     @Query("UPDATE prompt_history SET syncStatus = :newStatus WHERE syncStatus = :oldStatus")
     suspend fun updateAllSyncStatuses(oldStatus: String, newStatus: String)
 
     @Query("UPDATE prompt_history SET firestoreDocId = :docId WHERE id = :id")
-    suspend fun updateFirestoreDocId(id: Int, docId: String)
+    suspend fun updateFirestoreDocId(id: Long, docId: String)
 
     @Query("SELECT * FROM prompt_history WHERE id = :id")
-    suspend fun getPromptById(id: Int): PromptEntity?
+    suspend fun getPromptById(id: Long): PromptEntity?
 
     @Query("UPDATE prompt_history SET syncStatus = :newStatus WHERE id = :id AND text = :expectedText")
-    suspend fun markSyncedIfTextMatches(id: Int, expectedText: String, newStatus: String): Int
+    suspend fun markSyncedIfTextMatches(id: Long, expectedText: String, newStatus: String): Int
 }
